# IO-649 - Improve the performance of the contentEquals() methods.