--- a/src/changes/release-notes.vm
+++ b/src/changes/release-notes.vm
@@ -691,7 +691,7 @@ Enhancements from 1.2
   - New class designed for subclassing to walk through a set of files.
     DirectoryWalker provides the walk of the directories, filtering of
     directories and files, and cancellation support. The subclass must provide
-    the specific behaviour, such as text searching or image processing.
+    the specific behavior, such as text searching or image processing.
 
 - IOCase
   - New class/enumeration for case-sensitivity control
@@ -982,7 +982,7 @@ Enhancements from 1.0
     Constant for an empty array of File objects
 
 - ConditionalFileFilter - new interface  [30705]
-    Defines the behaviour of list based filters
+    Defines the behavior of list based filters
 
 - AndFileFilter, OrFileFilter  [30705]
     Now support a list of filters to and/or