… .travis.yml
… RELEASE-NOTES.txt
… pom.xml
… src/changes/changes.xml
… src/main/java/org/apache/commons/io/input/DemuxInputStream.java
… src/main/java/org/apache/commons/io/input/MessageDigestCalculatingInputStream.java
… src/main/java/org/apache/commons/io/input/ReversedLinesFileReader.java
… src/main/java/org/apache/commons/io/input/Tailer.java
… src/main/java/org/apache/commons/io/input/UnsynchronizedByteArrayInputStream.java
… src/main/java/org/apache/commons/io/input/XmlStreamReader.java
… src/main/java/org/apache/commons/io/input/buffer/CircularBufferInputStream.java
… src/main/java/org/apache/commons/io/input/buffer/CircularByteBuffer.java
… src/main/java/org/apache/commons/io/input/buffer/PeekableInputStream.java
… src/main/java/org/apache/commons/io/monitor/FileAlterationMonitor.java
… src/main/java/org/apache/commons/io/monitor/FileAlterationObserver.java
… src/main/java/org/apache/commons/io/output/AbstractByteArrayOutputStream.java
… src/main/java/org/apache/commons/io/output/ByteArrayOutputStream.java
… src/main/java/org/apache/commons/io/output/UnsynchronizedByteArrayOutputStream.java
… src/main/java/org/apache/commons/io/output/XmlStreamWriter.java
… src/site/xdoc/upgradeto1_1.xml
… src/site/xdoc/upgradeto1_3.xml
… src/site/xdoc/upgradeto2_5.xml
… src/test/java/org/apache/commons/io/FileUtilsCopyDirectoryToDirectoryTestCase.java
… src/test/java/org/apache/commons/io/IOUtilsTestCase.java
… src/test/java/org/apache/commons/io/SingleByteReadVsMultiByteReadPerformanceTest.java
… src/test/java/org/apache/commons/io/filefilter/FileFilterTestCase.java
… src/test/java/org/apache/commons/io/function/IOFunctionTest.java
… src/test/java/org/apache/commons/io/input/CharSequenceReaderTest.java
… src/test/java/org/apache/commons/io/input/ObservableInputStreamTest.java
… src/test/java/org/apache/commons/io/input/ReversedLinesFileReaderTestSimple.java