@brettlounsbury 2020-01-13
I took a look at the coverage drop.  I forgot to add tests that exercise the bounds checking of the bufferSize inputs to both contentEquals method.  I've pushed a new version with that.