@garydgregory 2020-01-24
Hi @brettlounsbury.

Thank you for your PR and updates. I've sprinkled comments throughout.

Performance claims must be backed up somehow, please add a JMH benchmark to the test so myself and others can compare mileage. 

For examples of JMH usage, please see other Apache Commons components like Commons BCEL, Commons Crypto, Commons CSV, Commons Lang, Commons Math, and Commons RNG.

Thank you!
Gary