@brettlounsbury 2020-01-27
@garydgregory,

I really appreciate your thorough review of my code.  I think at this point I have addressed your comments, but if you have any others please let me know.

I will add that I made the IOUtils.DEFAULT_BUFFER_SIZE constant package private (default visibility) so it could be referenced from IOUtils tests so that I can ensure the logic makes sense around buffer boundaries and I didn't want to clone that property since it could later be mutated and invalidate some of the edge condition testing being done.