@brettlounsbury 2020-03-02
> Sorry about the delay, last day of vacation here ;-)

No worries - hope you had a great vacation!