@brettlounsbury 2020-05-18
@garydgregory - resolved the merge conflict, should be good now.